Chapter4
Chapter 4 assignment from murach's HTML and CSS Training and Reference book.

2 Hours to complete though for a first go at HTML/CSS I'm not sure that's too bad. Struggled ALOT with getting the text colors to work with links, missing semi-colons, missing var().

Regardless the links are color coded, and they aren't hard coded that way so it a success even if a slow one.

The page isn't identical to the example they give but slightly different colors and font size are in my opinion forgivable. Also I fixed a typo so that should count as a bonus.