Chapter6
Completed the Chapter 5 and 6 assignments from murach's HTML and CSS Training and Reference Book.

I have replicated the site both visually and functionally though after consulting the code they use to generate the page I can say for sure that I am starting to have a very different looking back end. Mostly because of different formatting but they are now meaningful differences in the code and not just slightly different colors.

It took 3 hours this time though I was also doing other things at the time so lets call it 2.5 hours.

Most of my troubles came from trying to match subtle things like padding and margins, the site was practically the same in half that time but I wanted to to look pretty much identical.