Chapter 12
Added a number of pages that are "under construction" added responsive scaling to text, images, and some margins. Added a form page to sign up and become a member, added a navigation bar at the top of the screen, added a media query that moves the aside under the section when the screen width gets low enough.